:recycle: refactor: 함수제거 한 파일내에서 관리
